+++ Badluck-Achievements/Components/Models/SteamPlayerGame.cs
@@ -8,6 +8,6 @@
         public double playtimeForever { get; set; } = 0; // in hours
         public double? playtime2Weeks { get; set; } = 0; // in hours
         public string iconUrl { get; set; } = string.Empty;
-        public uint completedAchievements { get; set; } = 0;
+        public long completedAchievements { get; set; } = 0;
     }
 }
